Improve mem usage callback class
- Added several unit tests
