+++ pymlearn/cv_ts.py
@@ -4,7 +4,29 @@
 from pdb import set_trace as debug
 import copy
 import pandas as pd
-from pymlearn.cross_validate_utils import TimeSeriesSplitter
+
+
+class TimeSeriesSplitter(object):
+    def __init__(self, train_size, n_ahead=1):
+        self.train_size = train_size
+        self.n_ahead = n_ahead
+
+    def split(self, data):
+        """Split the data into a series of train and validate sets
+
+        :param data: Data to be split on
+        :return:
+        """
+        data.reset_index(drop=True, inplace=True)
+        length = len(data)
+        assert length >= self.train_size + self.n_ahead, 'Data size is too small'
+
+        for idx in range(self.train_size, length - self.n_ahead + 1):
+            train_rows = range(idx - self.train_size, idx)
+            val_rows = range(idx, idx + self.n_ahead)
+
+            yield data.loc[train_rows].reset_index(drop=True),\
+                  data.loc[val_rows].reset_index(drop=True)
 
 
 def ts_cross_validate(estimator, data,
